moved localstorage logic for move card operation
