+++ public/mover.js
@@ -1,7 +1,10 @@
+import Storage from "./storage.js";
+
 const MOVE_HERE_TEXT = "— Move here —";
 
 export default class Mover {
   constructor() {
+    this.storage = new Storage();
   }
 
   startMoving(card) {
@@ -18,7 +21,7 @@ export default class Mover {
     });
     let cards = document.querySelectorAll(".card"); 
     cards.forEach((element) => {
-      if (! element.classList.contains("template")) 
+      if (!element.classList.contains("template")) 
       {
         let cb_clone = move_button.cloneNode(true);
         element.parentNode.insertBefore(cb_clone, element);
@@ -29,7 +32,7 @@ export default class Mover {
     move_buttons.forEach ((element) => {
       element.addEventListener("click", (event) => {
         event.target;
-        this.updateLocalStorage(element);
+        this.storage.moveCard(element, card);
         
         element.parentNode.insertBefore(card, element);
       
@@ -37,8 +40,6 @@ export default class Mover {
         move_buttons.forEach((el)=>{
           el.remove(); 
         });
-        this.movedCard = null;
-
       });
     });
   }
@@ -54,65 +55,4 @@ export default class Mover {
     });
   }
 
-  updateLocalStorage(moveButton) {
-    let cardsLS = JSON.parse(localStorage.getItem("cards"));
-    let movedCardID = this.getCardIDFromLS(this.movedCard);
-    console.log(movedCardID);
-    console.log(cardsLS[movedCardID]);
-
-    let newColumn = moveButton.parentNode.id;
-
-    let numCardsBeforeMB = 0;
-    let prevCard;
-    for (let el of moveButton.parentNode.childNodes) 
-    {
-      if (el === moveButton) { break; }
-      if (el.classList !== undefined && el.classList.contains("card")) 
-      {
-        prevCard = el;
-        numCardsBeforeMB += 1;
-      }
-    }
-    if (numCardsBeforeMB === 0)
-    {
-      let cTemp = cardsLS.splice(movedCardID, 1)[0];
-      cardsLS.splice(0, 0, cTemp);
-      movedCardID = 0;
-    }
-    else 
-    {
-      let prevCardID = this.getCardIDFromLS(prevCard);
-      
-      let cTemp = cardsLS.splice(movedCardID, 1)[0];
-      cardsLS.splice(prevCardID, 0, cTemp);
-      movedCardID = prevCardID;
-    }
-    console.log(movedCardID);
-    console.log(cardsLS[movedCardID]);
-
-    cardsLS[movedCardID]["cardColumn"] = newColumn;
-    localStorage.setItem("cards", JSON.stringify(cardsLS));
-  }
-
-  getCardIDFromLS (card, column = card.parentNode.id) {
-    let cardsLS = JSON.parse(localStorage.getItem("cards"));
-    let cTitle = card.querySelector(".title").textContent;
-
-    //RGB to hex conversion from StackOverflow: 
-    //https://stackoverflow.com/questions/1740700/how-to-get-hex-color-value-rather-than-rgb-value
-    const rgba2hex = (rgba) => `#${rgba.match(/^rgba?\((\d+),\s*(\d+),\s*(\d+)(?:,\s*(\d+\.{0,1}\d*))?\)$/).slice(1).map((n, i) => (i === 3 ? Math.round(parseFloat(n) * 255) : parseFloat(n)).toString(16).padStart(2, '0').replace('NaN', '')).join('')}`
-    let cColor = rgba2hex(card.style.background);
-    
-    let cDescription;
-    if (card.querySelector(".description").textContent === "(No description)")
-      cDescription = null;
-    else 
-      cDescription = card.querySelector(".description").textContent; 
-      for (let c of cardsLS) {
-      if (cTitle === c["cardTitle"] && cColor === c["cardColor"] && column === c["cardColumn"] && cDescription === c["cardDescription"]) 
-      {
-        return cardsLS.indexOf(c);
-      }
-    }
-  }
 }
