+++ public/styles2.css
@@ -5,6 +5,8 @@
     --color-todo-done: #f0f0f0;
     --color-doing: white;
     --color-column-titles: black;
+    --color-card-bg: #ffffff;
+    --color-card-shadow: rgba(0, 0, 0, 0.446);
 }
 [data-theme="dark"] {
     --color-bg: #202020;
@@ -13,10 +15,12 @@
     --color-todo-done: #3b3b3b;
     --color-doing: #202020;
     --color-column-titles: white;
+    --color-card-bg: #333333;
+    --color-card-shadow: rgba(255, 255, 255, 0.473);
 }
 
 body {
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+    font-family: 'Inter';
     margin: 0;
     padding: 0;
     height: fit-content;
@@ -31,13 +35,13 @@ header {
     display: flex;
     justify-content: space-between;
     align-content: center;
-    padding: 0.5rem 5rem;
+    padding: 1.5rem 5rem;
     /* width: 100vw; */
     box-sizing: border-box;
 }
 h1 {
-    font-size: x-large;
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
+    font-size: xx-large;
+    /* font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif; */
     margin: 0;
 }
 
@@ -81,7 +85,7 @@ h1 {
 
 #board {
     /* width: 100%; */
-    height: 100%;
+    /* height: 100%; */
     min-height: 100vh;
     padding: 0 5rem;
     display: flex;
@@ -118,9 +122,11 @@ h1 {
     background-color: var(--color-doing);
 }
 .card {
-    border-radius: 0.5rem;
+    border-radius: 0.75rem;
     padding: 0.5rem;
     position: relative;
+    background-color: var(--color-card-bg);
+    box-shadow: 0 2px 4px var(--color-card-shadow);
 }
 .card h3 {
     margin: 0;
